docs static files served
