+++ airflow/opt/airflow.cfg
@@ -102,7 +102,7 @@ plugins_folder = /opt/airflow/plugins
 execute_tasks_new_python_interpreter = False
 
 # Secret key to save connection passwords in the db
-fernet_key = 
+fernet_key =
 
 # Whether to disable pickling dags
 donot_pickle = True
@@ -185,7 +185,7 @@ xcom_backend = etna.xcom.EtnaXCom
 
 # By default Airflow plugins are lazily-loaded (only loaded when required). Set it to ``False``,
 # if you want to load plugins whenever 'airflow' is invoked via cli or loaded from module.
-lazy_load_plugins = True
+lazy_load_plugins = False
 
 # By default Airflow providers are lazily-discovered (discovery and imports happen only when required).
 # Set it to False, if you want to discover providers whenever 'airflow' is invoked via cli or
@@ -472,7 +472,7 @@ worker_refresh_interval = 6000
 
 # If set to True, Airflow will track files in plugins_folder directory. When it detects changes,
 # then reload the gunicorn.
-reload_on_plugin_change = False
+reload_on_plugin_change = True
 
 # Secret key used to run your flask app. It should be as random as possible. However, when running
 # more than 1 instances of webserver, make sure all of them use the same ``secret_key`` otherwise
