+++ airflow/opt/plugins/etna/etna_docs_plugin.py
@@ -1,0 +1,35 @@
+from airflow.plugins_manager import AirflowPlugin
+from flask import Blueprint
+
+import logging
+
+
+ETNA_DOCS_MENU_TITLE = "Etna Airflow Docs"
+ETNA_DOCS_STATIC_PATH = "/static/docs"
+
+# Creating flask appbuilder Menu Items
+appbuilder_mitem_toplevel = {
+    "name": ETNA_DOCS_MENU_TITLE,
+    "href": f"{ETNA_DOCS_STATIC_PATH}/index.html",
+}
+
+# Creating a flask blueprint to integrate the doc static folder
+docs_blueprint = Blueprint(
+    "etna_plugin_docs_blueprint",
+    __name__,
+    template_folder='templates',
+    static_folder='static/docs',
+    static_url_path=ETNA_DOCS_STATIC_PATH)
+
+
+class EtnaDocsPlugin(AirflowPlugin):
+    name = "etna_docs_plugin"
+
+    appbuilder_menu_items = [appbuilder_mitem_toplevel]
+    flask_blueprints = [docs_blueprint]
+
+    def on_load(*args, **kwargs):
+        # ... perform Plugin boot actions
+        logging.error("here I have been loaded")
+        logging.error(args)
+        logging.error(docs_blueprint)
