+++ airflow/opt/plugins/etna/githook_plugin.py
@@ -1,0 +1,10 @@
+from airflow.plugins_manager import AirflowPlugin
+
+
+from etna.hooks.git import GitHook
+
+
+class GitHookPlugin(AirflowPlugin):
+    name = "githook_plugin"
+
+    hooks = [GitHook]
