+++ airflow/opt/plugins/etna/static/docs/etna/dags/project_name.html
