+++ airflow/opt/plugins/etna/static/docs/etna/operators/docker_operator_base.html
