+++ airflow/opt/plugins/etna/static/docs/etna/operators/swarm_operator.html
