+++ airflow/opt/plugins/etna/static/docs/index.html
