+++ airflow/opt/plugins/etna/static/docs/search.js
